% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dish.R
\name{dish_cleanup}
\alias{dish_cleanup}
\title{Clean up lazybtw dish sessions}
\usage{
dish_cleanup(all = FALSE, remove_dishes = TRUE)
}
\arguments{
\item{all}{Whether to remove all lazybtw session records. The default only
removes dead, corrupt, or unsafe sessions.}

\item{remove_dishes}{Whether to also remove the dish files for removed
sessions.}
}
\description{
Clean up lazybtw dish sessions
}
