% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dish.R
\name{dish_clear}
\alias{dish_clear}
\title{Clear the current lazybtw dish}
\usage{
dish_clear(confirm = interactive())
}
\arguments{
\item{confirm}{Whether to ask before clearing.}
}
\description{
Clear the current lazybtw dish
}
