% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw-package.R
\docType{package}
\name{lazybtw-package}
\alias{lazybtw}
\alias{lazybtw-package}
\title{lazybtw: Lightweight R Context Bridge for LLM Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Collects lightweight 'R' analysis context and exposes it through local workflows for Large Language Model (LLM) tools. Provides utilities for describing 'R' objects, package documentation, and workspace state in plain text formats optimized for LLM consumption. Includes a local dish for clipboard-free context capture and future read-only Model Context Protocol integrations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/posit-dev/btw}
  \item \url{https://posit-dev.github.io/btw/}
  \item Report bugs at \url{https://github.com/posit-dev/btw/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Authors:
\itemize{
  \item Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Google (Material Design Icons) [copyright holder]
  \item Microsoft (@vscode/codicons) [copyright holder]
  \item Jamie Perkins (countUp.js author) [copyright holder]
}

}
\keyword{internal}
